graph: set SVG definitions in respective component
SVG's defs tag allows SVG to reuse an element or define an element that
would be used as part of mask. This, in convential web used to be global
to the document but with real shadow DOM, it now is encapsulated well
inside a WebComponent. This became a problem for TensorBoard for folowing
reason.

tf-graph-scene defines all the reusable gradient/elements for other
components to use and with the real shadow DOM, tf-graph-controls and
node info module broke badly. Now, this is partially alleviated by
factoring out icon rendering into a shareable component that renders
all appropriate SVG elements. However, color is a bit trickier. Color,
though is only a handful of predefined set, currently is defined per
node and a color definition includes a name of the node. For this
reason, currently, we dynamically define the gradients for each nodes
as it is used in tf-graph-scene and this behavior needed to be changed
so we populate the gradients whereever it was being used.