Fix graph fitting in tf-graph-controls
tf-graph-controls was using global querySelector which does not
work as expected with WebComponents. This code was especially less ideal
since tf-graph-controls does not render tf-graph-scene.

Addressed the issue by bubbling up user's action to a component that
renders the tf-graph-scene and created public fit method for all
components that are in between tf-graph-dashboard and tf-graph-scene.