pr-curve: fix the basic functionalities
Problem 1: the DOM structure
Problem 2: tf-pr-curve-steps-selector broke because its usage of
querySelector. It uses an internal state, `runsToStepIndex`, to record
run to stepIndex (note the distinction from just step)_map. It was
manually setting the value to the paper-sliders and listened to the
changed event via listeners. Something in Polymer 2 broke. I took this
opportunity to refactor to leverage the two-way binding.

stepIndex vs. step: when there are tons of data, due to our reservoir
sampling, there can be only a subset of the steps. Say we have steps
[0, 4, 10, 1000] for a run foo. When the steps are rendered, it has to
translated to indices since the paper-slider does not supporting an enum.
`runToStepIndex` stores slider index which gets translated to real
`step` via `runToAvailableTimeEntries`.